refactor: rename widget folder MorphoVaultWidget to VaultWidget
The widget serves both Morpho and Spark via provider gating. git mv the folder and updated all import paths (widgets barrel, the shared interactive stats card, the internal test mock path) plus the ErrorBoundary label. The MorphoVaultWidget symbol alias is unchanged.
