+++ apps/webapp/src/widgets/VaultWidget/components/SparkVaultRate.tsx
