+++ apps/webapp/src/widgets/VaultWidget/components/VaultPoweredBy.tsx
