+++ apps/webapp/src/widgets/VaultWidget/components/sparkVaultRate.spark.test.tsx
